separate out heuristics test from main tests
