+++ sympy/solvers/ode/tests/test_ode.py
@@ -817,30 +817,6 @@ def test_homogeneous_order():
     raises(ValueError, lambda: homogeneous_order(x*y))
 
 
-def test_nth_linear_constant_coeff_homogeneous():
-    # Issue #15237
-    eqn = Derivative(x*f(x), x, x, x)
-    hint = 'nth_linear_constant_coeff_homogeneous'
-    raises(ValueError, lambda: dsolve(eqn, f(x), hint, prep=True))
-    raises(ValueError, lambda: dsolve(eqn, f(x), hint, prep=False))
-
-
-@XFAIL
-@slow
-def test_nth_linear_constant_coeff_homogeneous_rootof_sol():
-    # See https://github.com/sympy/sympy/issues/15753
-    if ON_TRAVIS:
-        skip("Too slow for travis.")
-    eq = f(x).diff(x, 5) + 11*f(x).diff(x) - 2*f(x)
-    sol = Eq(f(x),
-        C1*exp(x*rootof(x**5 + 11*x - 2, 0)) +
-        C2*exp(x*rootof(x**5 + 11*x - 2, 1)) +
-        C3*exp(x*rootof(x**5 + 11*x - 2, 2)) +
-        C4*exp(x*rootof(x**5 + 11*x - 2, 3)) +
-        C5*exp(x*rootof(x**5 + 11*x - 2, 4)))
-    assert checkodesol(eq, sol, order=5, solve_for_func=False)[0]
-
-
 @XFAIL
 def test_noncircularized_real_imaginary_parts():
     # If this passes, lines numbered 3878-3882 (at the time of this commit)
@@ -990,15 +966,6 @@ def test_issue_12623():
     #issue-https://github.com/sympy/sympy/issues/12623
 
 
-def test_unexpanded_Liouville_ODE():
-    # This is the same as eq1 from test_Liouville_ODE() above.
-    eq1 = diff(f(x), x)/x + diff(f(x), x, x)/2 - diff(f(x), x)**2/2
-    eq2 = eq1*exp(-f(x))/exp(f(x))
-    sol2 = Eq(f(x), C1 + log(x) - log(C2 + x))
-    sol2s = constant_renumber(sol2)
-    assert dsolve(eq2) in (sol2, sol2s)
-    assert checkodesol(eq2, sol2, order=2, solve_for_func=False)[0]
-
 def test_issue_4785():
     from sympy.abc import A
     eq = x + A*(x + diff(f(x), x) + f(x)) + diff(f(x), x) + f(x) + 2
@@ -1140,14 +1107,6 @@ def test_linear_coeff_match():
     assert _linear_coeff_match(eq7, f(x)) is None
 
 
-def test_linear_coefficients():
-    f = Function('f')
-    sol = Eq(f(x), C1/(x**2 + 6*x + 9) - Rational(3, 2))
-    eq = f(x).diff(x) + (3 + 2*f(x))/(x + 3)
-    assert dsolve(eq, hint='linear_coefficients') == sol
-    assert checkodesol(eq, sol, order=1, solve_for_func=False)[0]
-
-
 def test_constantsimp_take_problem():
     c = exp(C1) + 2
     assert len(Poly(constantsimp(exp(C1) + c + c*x, [C1])).gens) == 2
@@ -1184,44 +1143,6 @@ def test_issue_6989():
     assert checkodesol(eq, sol, order=1, solve_for_func=False)[0]
 
 
-def test_heuristic1():
-    y, a, b, c, a4, a3, a2, a1, a0 = symbols("y a b c a4 a3 a2 a1 a0")
-    f = Function('f')
-    xi = Function('xi')
-    eta = Function('eta')
-    df = f(x).diff(x)
-    eq = Eq(df, x**2*f(x))
-    eq1 = f(x).diff(x) + a*f(x) - c*exp(b*x)
-    eq2 = f(x).diff(x) + 2*x*f(x) - x*exp(-x**2)
-    eq3 = (1 + 2*x)*df + 2 - 4*exp(-f(x))
-    eq4 = f(x).diff(x) - (a4*x**4 + a3*x**3 + a2*x**2 + a1*x + a0)**Rational(-1, 2)
-    eq5 = x**2*df - f(x) + x**2*exp(x - (1/x))
-    eqlist = [eq, eq1, eq2, eq3, eq4, eq5]
-
-    i = infinitesimals(eq, hint='abaco1_simple')
-    assert i == [{eta(x, f(x)): exp(x**3/3), xi(x, f(x)): 0},
-        {eta(x, f(x)): f(x), xi(x, f(x)): 0},
-        {eta(x, f(x)): 0, xi(x, f(x)): x**(-2)}]
-    i1 = infinitesimals(eq1, hint='abaco1_simple')
-    assert i1 == [{eta(x, f(x)): exp(-a*x), xi(x, f(x)): 0}]
-    i2 = infinitesimals(eq2, hint='abaco1_simple')
-    assert i2 == [{eta(x, f(x)): exp(-x**2), xi(x, f(x)): 0}]
-    i3 = infinitesimals(eq3, hint='abaco1_simple')
-    assert i3 == [{eta(x, f(x)): 0, xi(x, f(x)): 2*x + 1},
-        {eta(x, f(x)): 0, xi(x, f(x)): 1/(exp(f(x)) - 2)}]
-    i4 = infinitesimals(eq4, hint='abaco1_simple')
-    assert i4 == [{eta(x, f(x)): 1, xi(x, f(x)): 0},
-        {eta(x, f(x)): 0,
-        xi(x, f(x)): sqrt(a0 + a1*x + a2*x**2 + a3*x**3 + a4*x**4)}]
-    i5 = infinitesimals(eq5, hint='abaco1_simple')
-    assert i5 == [{xi(x, f(x)): 0, eta(x, f(x)): exp(-1/x)}]
-
-    ilist = [i, i1, i2, i3, i4, i5]
-    for eq, i in (zip(eqlist, ilist)):
-        check = checkinfsol(eq, i)
-        assert check[0]
-
-
 def test_issue_6247():
     eq = f(x).diff(x, x) + 4*f(x)
     sol = Eq(f(x), C1*sin(2*x) + C2*cos(2*x))
@@ -1229,106 +1150,18 @@ def test_issue_6247():
     assert checkodesol(eq, sol, order=1)[0]
 
 
-def test_heuristic2():
-    xi = Function('xi')
-    eta = Function('eta')
-    df = f(x).diff(x)
-
-    # This ODE can be solved by the Lie Group method, when there are
-    # better assumptions
-    eq = df - (f(x)/x)*(x*log(x**2/f(x)) + 2)
-    i = infinitesimals(eq, hint='abaco1_product')
-    assert i == [{eta(x, f(x)): f(x)*exp(-x), xi(x, f(x)): 0}]
-    assert checkinfsol(eq, i)[0]
 
 
-@slow
-def test_heuristic3():
-    xi = Function('xi')
-    eta = Function('eta')
-    a, b = symbols("a b")
-    df = f(x).diff(x)
-
-    eq = x**2*df + x*f(x) + f(x)**2 + x**2
-    i = infinitesimals(eq, hint='bivariate')
-    assert i == [{eta(x, f(x)): f(x), xi(x, f(x)): x}]
-    assert checkinfsol(eq, i)[0]
-
-    eq = x**2*(-f(x)**2 + df)- a*x**2*f(x) + 2 - a*x
-    i = infinitesimals(eq, hint='bivariate')
-    assert checkinfsol(eq, i)[0]
 
 
-def test_heuristic_4():
-    y, a = symbols("y a")
-
-    eq = x*(f(x).diff(x)) + 1 - f(x)**2
-    i = infinitesimals(eq, hint='chi')
-    assert checkinfsol(eq, i)[0]
 
 
-def test_heuristic_function_sum():
-    xi = Function('xi')
-    eta = Function('eta')
-    eq = f(x).diff(x) - (3*(1 + x**2/f(x)**2)*atan(f(x)/x) + (1 - 2*f(x))/x +
-       (1 - 3*f(x))*(x/f(x)**2))
-    i = infinitesimals(eq, hint='function_sum')
-    assert i == [{eta(x, f(x)): f(x)**(-2) + x**(-2), xi(x, f(x)): 0}]
-    assert checkinfsol(eq, i)[0]
 
 
-def test_heuristic_abaco2_similar():
-    xi = Function('xi')
-    eta = Function('eta')
-    F = Function('F')
-    a, b = symbols("a b")
-    eq = f(x).diff(x) - F(a*x + b*f(x))
-    i = infinitesimals(eq, hint='abaco2_similar')
-    assert i == [{eta(x, f(x)): -a/b, xi(x, f(x)): 1}]
-    assert checkinfsol(eq, i)[0]
-
-    eq = f(x).diff(x) - (f(x)**2 / (sin(f(x) - x) - x**2 + 2*x*f(x)))
-    i = infinitesimals(eq, hint='abaco2_similar')
-    assert i == [{eta(x, f(x)): f(x)**2, xi(x, f(x)): f(x)**2}]
-    assert checkinfsol(eq, i)[0]
 
 
-def test_heuristic_abaco2_unique_unknown():
-    xi = Function('xi')
-    eta = Function('eta')
-    F = Function('F')
-    a, b = symbols("a b")
-    x = Symbol("x", positive=True)
-
-    eq = f(x).diff(x) - x**(a - 1)*(f(x)**(1 - b))*F(x**a/a + f(x)**b/b)
-    i = infinitesimals(eq, hint='abaco2_unique_unknown')
-    assert i == [{eta(x, f(x)): -f(x)*f(x)**(-b), xi(x, f(x)): x*x**(-a)}]
-    assert checkinfsol(eq, i)[0]
-
-    eq = f(x).diff(x) + tan(F(x**2 + f(x)**2) + atan(x/f(x)))
-    i = infinitesimals(eq, hint='abaco2_unique_unknown')
-    assert i == [{eta(x, f(x)): x, xi(x, f(x)): -f(x)}]
-    assert checkinfsol(eq, i)[0]
-
-    eq = (x*f(x).diff(x) + f(x) + 2*x)**2 -4*x*f(x) -4*x**2 -4*a
-    i = infinitesimals(eq, hint='abaco2_unique_unknown')
-    assert checkinfsol(eq, i)[0]
 
 
-def test_heuristic_linear():
-    a, b, m, n = symbols("a b m n")
-
-    eq = x**(n*(m + 1) - m)*(f(x).diff(x)) - a*f(x)**n -b*x**(n*(m + 1))
-    i = infinitesimals(eq, hint='linear')
-    assert checkinfsol(eq, i)[0]
-
-
-@XFAIL
-def test_kamke():
-    a, b, alpha, c = symbols("a b alpha c")
-    eq = x**2*(a*f(x)**2+(f(x).diff(x))) + b*x**alpha + c
-    i = infinitesimals(eq, hint='sum_function')  # XFAIL
-    assert checkinfsol(eq, i)[0]
 
 
 def test_series():
@@ -1355,15 +1188,6 @@ def test_series():
 def test_lie_group_issue15219():
     eqn = exp(f(x).diff(x)-f(x))
     assert 'lie_group' not in classify_ode(eqn, f(x))
-
-
-def test_user_infinitesimals():
-    x = Symbol("x") # assuming x is real generates an error
-    eq = x*(f(x).diff(x)) + 1 - f(x)**2
-    sol = Eq(f(x), (C1 + x**2)/(C1 - x**2))
-    infinitesimals = {'xi':sqrt(f(x) - 1)/sqrt(f(x) + 1), 'eta':0}
-    assert dsolve(eq, hint='lie_group', **infinitesimals) == sol
-    assert checkodesol(eq, sol) == (True, 0)
 
 
 def test_issue_7081():
